+++ arduino/firmware/include/gpio.h
@@ -30,27 +30,27 @@ enum class IOType { DIGITAL_INPUT, DIGITAL_OUTPUT, ANALOG_INPUT, ANALOG_OUTPUT, 
 class GPIO
 {
 public:
-    GPIO(short pinNumber, IOType ioType);
+    GPIO(int pinNumber, IOType ioType);
     bool g_digitalRead();
     bool g_softDigitalRead();
     void g_digitalWrite(bool logicState);
-    short g_analogRead();
-    short g_softAnalogRead();
-    void g_analogWrite(short state);
+    int g_analogRead();
+    int g_softAnalogRead();
+    void g_analogWrite(int state);
     
     IOType ioType() const;
-    short pinNumber() const;
+    int pinNumber() const;
     
     void setIOType(IOType ioType);
-    void setPinNumber(short pinNumber);
+    void setPinNumber(int pinNumber);
 
-    short getIOAgnosticState();
+    int getIOAgnosticState();
     std::vector<unsigned char> getEEPROMWritableState();
-    static std::vector<unsigned char> toEEPROMWritableState(short longState);
+    static std::vector<unsigned char> toEEPROMWritableState(int longState);
 
-    static const short ANALOG_MAX;
-    static void setAnalogToDigitalThreshold(short threshold);
-    static short analogToDigitalThreshold();
+    static const int ANALOG_MAX;
+    static void setAnalogToDigitalThreshold(int threshold);
+    static int analogToDigitalThreshold();
 
     friend bool operator==(const GPIO &lhs, const GPIO &rhs)
     {
@@ -64,12 +64,12 @@ public:
     }
 
 private:
-    short m_pinNumber;
+    int m_pinNumber;
     IOType m_ioType;
     bool m_logicState;
-    short m_analogState;
+    int m_analogState;
 
-    static short s_analogToDigitalThreshold;
+    static int s_analogToDigitalThreshold;
 };
 
 #endif //ARDUINOPC_GPIO_H
